third_party/blink/web_tests/fast/css/font-face-unicode-range-load-expected.txt

Tests that unicode-range descriptor is used to selectively download fonts that are used in document.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS latin1Loaded is false
PASS cyrillicLoaded is true
PASS arabicLoaded is false
PASS iconLoaded is false
PASS successfullyParsed is true

TEST COMPLETE


Проверка

hello
